@@ -2,6 +2,7 @@
 	"name": "next-app",
 	"version": "0.1.0",
 	"private": true,
+	"type": "module",
 	"scripts": {
 		"dev": "run-script-os",
 		"dev:win32": "npx prisma generate && set NODE_OPTIONS='--inspect' && next dev",
@@ -21,6 +22,7 @@
 		"@prisma/client": "^5.9.0",
 		"autoprefixer": "^10.4.14",
 		"crypto-js": "^4.2.0",
+		"esbuild-register": "^3.5.0",
 		"gray-matter": "^4.0.3",
 		"html-react-parser": "^2.0.0",
 		"next": "^14.1.0",
@@ -37,7 +39,8 @@
 		"remark": "^14.0.2",
 		"remark-gfm": "^3.0.1",
 		"remark-html": "^15.0.1",
-		"sass": "^1.53.0"
+		"sass": "^1.53.0",
+		"ts-node": "^10.9.2"
 	},
 	"devDependencies": {
 		"@chakra-ui/react": "^2.2.1",
@@ -53,7 +56,7 @@
 		"run-script-os": "^1.1.6",
 		"serve": "^14.2.0",
 		"tailwindcss": "^3.3.3",
-		"typescript": "^4.7.4"
+		"typescript": "^4.9.5"
 	},
 	"main": "next.config.js",
 	"repository": {
