@@ -1,20 +1,19 @@
-const axios = require("axios")
+import axios from "axios"
 
-module.exports = {
-	github: axios.create({
-		baseURL: "https://api.github.com",
-	}),
-	repos: [
-		"Haine",
-		"Youto",
-		"thumbelina-shell",
-		"kirinokougai",
-		"GhostSpeaker",
-		"Ukaing",
-		"GhostWardrobe",
-		"Bouyomi",
-		"recentghosts",
-		"shioriupdater",
-		"sunset-sunrise-saori",
-	],
-}
+export const github = axios.create({
+  baseURL: "https://api.github.com",
+})
+
+export const repos = [
+  // "Haine",
+  "Youto",
+  // "thumbelina-shell",
+  // "kirinokougai",
+  // "GhostSpeaker",
+  // "Ukaing",
+  // "GhostWardrobe",
+  // "Bouyomi",
+  // "recentghosts",
+  // "shioriupdater",
+  // "sunset-sunrise-saori",
+]
